fixing some stuff

+++ AVLTreeDebug.cpp
@@ -81,7 +81,7 @@ int main() {
     std::cout << std::endl;
     //
 //    // findRange
-    vector<int> rangeTest = tree.findRange("D", "W");
+    vector<size_t> rangeTest = tree.findRange("D", "W");
     // 70 68 82 75 77 86
     for (auto val: rangeTest) {
         cout << val << " ";
